+++ lib/preload_google_ads.dart
@@ -13,6 +13,7 @@ export 'src/ad_commons/constants.dart';
 export 'src/ad_commons/preload_data_model.dart';
 export 'src/ad_counter/ad_counter.dart';
 export 'src/ad_initial.dart';
+export 'src/app_open_ad/on_demand_app_open_ad.dart';
 export 'src/banner_ad/show_collapsible_banner_ad.dart';
 export 'src/inter_ad/on_demand_interstitial_ad.dart';
 export 'src/native_ad/native_ad_style.dart';